macOS: Disable threaded GL rendering on SDK 10.14+
AppKit expects rendering to happen on the main thread, or at least any
interaction with AppKit UI classes such as NSView. Our OpenGL helpers,
such as QOpenGLContext, do not enforce this, and we may end up calling
into AppKit UI classes on the render thread, deadlocking the application.

Until this can be investigated and new APIs possibly introduced that allow
a more fine grained control in our own classes, we disable threaded GL
as a capability of the platform, which will inform clients such as
QtQuick to use the basic render loop.

[ChangeLog][macOS] Threaded OpenGL usage has been disabled when building
using Xcode 10/SDK 10.14 and later. Qt Quick defaults to the 'basic' render
loop now on macOS.

Task-number: QTBUG-71731
Change-Id: I6fc3295e833ecd48ad49382b8275c762fa7978a6
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>