QWindowsLocalCodec::convertToUnicode(): Fix remaining char conversion
Prevent passing a zero argument to the cbMultiByte parameter of
MultiByteToWideChar. According to the documentation this will always
fail.

This is triggered when calling convertToUnicode with a state having
remainingChars of one and a length of one. The remaining chars will be
completed and the length will be reduced to zero before the actual
convert.

Adding an early return after the completion to prevent calling
MultiByteToWideChar with the invalid zero argument.

Fixes: QTCREATORBUG-21622
Change-Id: Ic6a433dd929e97b1681583cb1b9a347afaa9a09f
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>